begin decomposing Fixpoint.Types
