+++ src/Language/Fixpoint/Types.hs
@@ -200,10 +200,6 @@ import qualified Data.HashMap.Strict       as M
 import qualified Data.HashSet              as S
 
 
-traceFix     ::  (Fixpoint a) => String -> a -> a
-traceFix s x = trace ("\nTrace: [" ++ s ++ "] : " ++ showFix x) x
-
-type TCEmb a    = M.HashMap a FTycon
 
 exprSymbols :: Expr -> [Symbol]
 exprSymbols = go
@@ -263,110 +259,6 @@ ksMember k (KS s) = S.member k s
 --------------------------------------------------------------------------------
 -- | Type Constructors ---------------------------------------------------------
 --------------------------------------------------------------------------------
-
-newtype FTycon = TC LocSymbol deriving (Eq, Ord, Show, Data, Typeable, Generic)
-
-intFTyCon, boolFTyCon, realFTyCon, funcFTyCon, numFTyCon, strFTyCon, propFTyCon, listFTyCon :: FTycon
-intFTyCon  = TC $ dummyLoc "int"
-boolFTyCon = TC $ dummyLoc "bool"
-realFTyCon = TC $ dummyLoc "real"
-numFTyCon  = TC $ dummyLoc "num"
-funcFTyCon = TC $ dummyLoc "function"
-strFTyCon  = TC $ dummyLoc strConName
-propFTyCon = TC $ dummyLoc propConName
-listFTyCon = TC $ dummyLoc listConName
-
-isListConName :: LocSymbol -> Bool
-isListConName x = c == listConName || c == listLConName --"List"
-  where
-    c           = val x
-
-isListTC :: FTycon -> Bool
-isListTC (TC z) = isListConName z
-
-fTyconSymbol :: FTycon -> Located Symbol
-fTyconSymbol (TC s) = s
-
-symbolFTycon :: LocSymbol -> FTycon
-symbolFTycon c
-  | isListConName c
-  = TC $ fmap (const listConName) c
-  | otherwise
-  = TC c
-
-fApp :: Sort -> [Sort] -> Sort
-fApp = foldl' FApp
-
-fAppTC :: FTycon -> [Sort] -> Sort
-fAppTC = fApp . fTyconSort
-
-fApp' :: Sort -> ListNE Sort
-fApp' = go []
-  where
-    go acc (FApp t1 t2) = go (t2 : acc) t1
-    go acc t            = t : acc
-
-fObj :: LocSymbol -> Sort
-fObj = fTyconSort . TC
-
-sortFTycon :: Sort -> Maybe FTycon
-sortFTycon FInt    = Just intFTyCon
-sortFTycon FReal   = Just realFTyCon
-sortFTycon FNum    = Just numFTyCon
-sortFTycon (FTC c) = Just c
-sortFTycon _       = Nothing
-
-----------------------------------------------------------------------
-------------------------------- Sorts --------------------------------
-----------------------------------------------------------------------
-
-data Sort = FInt
-          | FReal
-          | FNum                 -- ^ numeric kind for Num tyvars
-          | FFrac                -- ^ numeric kind for Fractional tyvars
-          | FObj  Symbol         -- ^ uninterpreted type
-          | FVar  !Int           -- ^ fixpoint type variable
-          | FFunc !Int ![Sort]   -- ^ type-var arity, in-ts ++ [out-t]
-          | FTC   FTycon
-          | FApp  Sort Sort      -- ^ constructed type
-              deriving (Eq, Ord, Show, Data, Typeable, Generic)
-
-{-@ FFunc :: Nat -> ListNE Sort -> Sort @-}
-
-instance Hashable Sort
-
-newtype Sub = Sub [(Int, Sort)] deriving (Generic)
-
-instance Fixpoint Sort where
-  toFix = toFixSort
-
-toFixSort :: Sort -> Doc
-toFixSort (FVar i)     = text "@"   <> parens (toFix i)
-toFixSort FInt         = text "int"
-toFixSort FReal        = text "real"
-toFixSort FFrac        = text "frac"
-toFixSort (FObj x)     = toFix x
-toFixSort FNum         = text "num"
-toFixSort (FFunc n ts) = text "func" <> parens (toFix n <> text ", " <> toFix ts)
-toFixSort (FTC c)      = toFix c
-toFixSort t@(FApp _ _) = toFixFApp (fApp' t)
-
-toFixFApp            :: ListNE Sort -> Doc
-toFixFApp [t]        = toFixSort t
-toFixFApp [FTC c, t]
-  | isListTC c       = brackets $ toFixSort t
-toFixFApp ts         = parens $ intersperse space (toFixSort <$> ts)
-
-instance Fixpoint FTycon where
-  toFix (TC s)       = toFix s
-
-------------------------------------------------------------------------
-sortSubst                  :: M.HashMap Symbol Sort -> Sort -> Sort
-------------------------------------------------------------------------
-sortSubst θ t@(FObj x)   = fromMaybe t (M.lookup x θ)
-sortSubst θ (FFunc n ts) = FFunc n (sortSubst θ <$> ts)
-sortSubst θ (FApp t1 t2) = FApp (sortSubst θ t1) (sortSubst θ t2)
-sortSubst _  t           = t
 
 
 instance Show Subst where
@@ -1718,129 +1610,18 @@ editDistance xs ys = table ! (m,n)
     dist (i,j) = minimum [table ! (i-1,j) + 1, table ! (i,j-1) + 1,
         if x ! i == y ! j then table ! (i-1,j-1) else 1 + table ! (i-1,j-1)]
 
-
 -----------------------------------------------------------------------------
 -- | Located Values ---------------------------------------------------------
 -----------------------------------------------------------------------------
 
-instance (IsString a) => IsString (Located a) where
-  fromString = dummyLoc . fromString
-
-type LocSymbol = Located Symbol
-type LocText   = Located Text
-
-
-locAt :: String -> a -> Located a
-locAt s  = Loc l l
-  where
-    l    = dummyPos s
-
-dummyLoc :: a -> Located a
-dummyLoc = Loc l l
-  where
-    l    = dummyPos "Fixpoint.Types.dummyLoc"
-
-dummyPos   :: String -> SourcePos
-dummyPos s = newPos s 0 0
-
-isDummy :: (Symbolic a) => a -> Bool
-isDummy a = symbol a == symbol dummyName
-
-instance Fixpoint SourcePos where
-  toFix = text . show
-
-instance Fixpoint a => Fixpoint (Located a) where
-  toFix = toFix . val
-
-instance Symbolic a => Symbolic (Located a) where
-  symbol = symbol . val
-
 instance Expression a => Expression (Located a) where
   expr   = expr . val
-
-instance Functor Located where
-  fmap f (Loc l l' x) =  Loc l l' (f x)
-
-
-instance Foldable Located where
-  foldMap f (Loc _ _ x) = f x
-
-instance Traversable Located where
-  traverse f (Loc l l' x) = Loc l l' <$> f x
-
-instance Show a => Show (Located a) where
-  show (Loc l l' x) = show x ++ " defined from: " ++ show l ++ " to: " ++ show l'
-
-instance Eq a => Eq (Located a) where
-  (Loc _ _ x) == (Loc _ _ y) = x == y
-
-instance Ord a => Ord (Located a) where
-  compare x y = compare (val x) (val y)
 
 instance Subable a => Subable (Located a) where
   syms (Loc _ _ x)   = syms x
   substa f (Loc l l' x) = Loc l l' (substa f x)
   substf f (Loc l l' x) = Loc l l' (substf f x)
   subst su (Loc l l' x) = Loc l l' (subst su x)
-
-instance Hashable a => Hashable (Located a) where
-  hashWithSalt i = hashWithSalt i . val
-
-
--------------------------------------------------------------------------
--- | Exported Basic Sorts -----------------------------------------------
--------------------------------------------------------------------------
-
-numSort, boolSort, intSort, propSort, realSort, strSort, funcSort :: Sort
-boolSort = fTyconSort boolFTyCon
-strSort  = fTyconSort strFTyCon
-intSort  = fTyconSort intFTyCon
-realSort = fTyconSort realFTyCon
-propSort = fTyconSort propFTyCon
-numSort  = fTyconSort numFTyCon
-funcSort = fTyconSort funcFTyCon
-
-fTyconSort :: FTycon -> Sort
-fTyconSort c
-  | c == intFTyCon  = FInt
-  | c == realFTyCon = FReal
-  | c == numFTyCon  = FNum
-  | otherwise       = FTC c
-
--------------------------------------------------------------------------
--- | Constraint Partition Container -------------------------------------
--------------------------------------------------------------------------
-
-
-data CPart a = CPart { pws :: M.HashMap KVar (WfC a)
-                     , pcm :: M.HashMap Integer (SubC a)
-                     , cFileName :: FilePath
-                     }
-
-instance Monoid (CPart a) where
-   mempty = CPart mempty mempty mempty
-   mappend l r = CPart { pws = pws l `mappend` pws r
-                       , pcm = pcm l `mappend` pcm r
-                       , cFileName = cFileName l
-                       }
-
--------------------------------------------------------------------------
--- | Multicore info -----------------------------------------------------
--------------------------------------------------------------------------
-
-data MCInfo = MCInfo { mcCores :: Int
-                     , mcMinPartSize :: Int
-                     , mcMaxPartSize :: Int
-                     } deriving (Show)
-
-mcInfo :: Config -> IO MCInfo
-mcInfo c = do
-   np <- getNumProcessors
-   let nc = fromMaybe np (cores c)
-   return MCInfo { mcCores = nc
-                 , mcMinPartSize = minPartSize c
-                 , mcMaxPartSize = maxPartSize c
-                 }
 
 ---------------------------------------------------------------------------
 -- | FInfo to SInfo conversion
